requirements.txt
numpy==1.23.5
pandas==2.0.2
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
mysql-connector-python==8.0.33
pymongo==4.4.1